version cmi aliyun

+++ url_crawler/url_crawler/start_crawler.sh
@@ -1,12 +1,15 @@
 #!/bin/bash
 while true
 do
-  file_data=$(cat signal.txt)
-  if [ "$file_data" = "over" ];then
-     killall -9 python
-     sleep 180
-     python start.py
-  elif [ "$file_data" = "start" ];then
-     sleep 10
+  hour=`date +%H`
+  if [ $hour -gt 7 -a $hour -lt 24 ];then
+      file_data=$(cat signal.txt)
+      if [ "$file_data" = "over" ];then
+         killall -9 python
+         sleep 3000
+         python start.py
+      elif [ "$file_data" = "start" ];then
+         sleep 20
+      fi
   fi
 done
